Update with failed commit leftovers
I was trying to see if sorting at the end would be faster than keeping
it sorted, but since we have to go through the list to see if a word has
been found anyway, it actually makes it an order of magnitude slower to
sort it later. However, in the process of doing this, I added some
changes that I figured I might as well leave in.